Moved Scenario from manager to basic_scenario
