+++ srunner/scenarios/basic_scenario.py
@@ -6,7 +6,7 @@
 # For a copy, see <https://opensource.org/licenses/MIT>.
 
 """
-This module provide the basic class for all user-defined scenarios.
+This module provide BasicScenario, the basic class of all the scenarios.
 """
 
 from __future__ import print_function
@@ -17,7 +17,8 @@ import carla
 
 import srunner.scenariomanager.scenarioatomics.atomic_trigger_conditions as conditions
 from srunner.scenariomanager.carla_data_provider import CarlaDataProvider
-from srunner.scenariomanager.scenario_manager import Scenario
+from srunner.scenariomanager.timer import TimeOut
+from srunner.scenariomanager.weather_sim import WeatherBehavior
 
 
 class BasicScenario(object):
@@ -202,3 +203,91 @@ class BasicScenario(object):
                     CarlaDataProvider.remove_actor_by_id(self.other_actors[i].id)
                 self.other_actors[i] = None
         self.other_actors = []
+
+
+class Scenario(object):
+
+    """
+    Basic scenario class. This class holds the behavior_tree describing the
+    scenario and the test criteria.
+
+    The user must not modify this class.
+
+    Important parameters:
+    - behavior: User defined scenario with py_tree
+    - criteria_list: List of user defined test criteria with py_tree
+    - timeout (default = 60s): Timeout of the scenario in seconds
+    - terminate_on_failure: Terminate scenario on first failure
+    """
+
+    def __init__(self, behavior, criteria, name, timeout=60, terminate_on_failure=False):
+        self.behavior = behavior
+        self.test_criteria = criteria
+        self.timeout = timeout
+        self.name = name
+
+        if self.test_criteria is not None and not isinstance(self.test_criteria, py_trees.composites.Parallel):
+            # list of nodes
+            for criterion in self.test_criteria:
+                criterion.terminate_on_failure = terminate_on_failure
+
+            # Create py_tree for test criteria
+            self.criteria_tree = py_trees.composites.Parallel(
+                name="Test Criteria",
+                policy=py_trees.common.ParallelPolicy.SUCCESS_ON_ONE
+            )
+            self.criteria_tree.add_children(self.test_criteria)
+            self.criteria_tree.setup(timeout=1)
+        else:
+            self.criteria_tree = criteria
+
+        # Create node for timeout
+        self.timeout_node = TimeOut(self.timeout, name="TimeOut")
+
+        # Create overall py_tree
+        self.scenario_tree = py_trees.composites.Parallel(name, policy=py_trees.common.ParallelPolicy.SUCCESS_ON_ONE)
+        if behavior is not None:
+            self.scenario_tree.add_child(self.behavior)
+        self.scenario_tree.add_child(self.timeout_node)
+        self.scenario_tree.add_child(WeatherBehavior())
+        if criteria is not None:
+            self.scenario_tree.add_child(self.criteria_tree)
+        self.scenario_tree.setup(timeout=1)
+
+    def _extract_nodes_from_tree(self, tree):  # pylint: disable=no-self-use
+        """
+        Returns the list of all nodes from the given tree
+        """
+        node_list = [tree]
+        more_nodes_exist = True
+        while more_nodes_exist:
+            more_nodes_exist = False
+            for node in node_list:
+                if node.children:
+                    node_list.remove(node)
+                    more_nodes_exist = True
+                    for child in node.children:
+                        node_list.append(child)
+
+        if len(node_list) == 1 and isinstance(node_list[0], py_trees.composites.Parallel):
+            return []
+
+        return node_list
+
+    def get_criteria(self):
+        """
+        Return the list of test criteria (all leave nodes)
+        """
+        criteria_list = self._extract_nodes_from_tree(self.criteria_tree)
+        return criteria_list
+
+    def terminate(self):
+        """
+        This function sets the status of all leaves in the scenario tree to INVALID
+        """
+        # Get list of all nodes in the tree
+        node_list = self._extract_nodes_from_tree(self.scenario_tree)
+
+        # Set status to INVALID
+        for node in node_list:
+            node.terminate(py_trees.common.Status.INVALID)
